requirements.txt
Flask
Flask-Cors
tensorflow
h5py
scipy
numpy
matplotlib
gunicorn 
google-cloud-firestore
python-dotenv
google-cloud-storage 